refactor: Rename People to Employees
separate form into own file.
