+++ apps/front/common/constants.tsx
@@ -1,6 +1,6 @@
 import { HiOutlineUsers, HiOutlineCurrencyDollar } from 'react-icons/hi2'
 
 export const DASHBOARD_TABS = [
-  { title: 'People', icon: <HiOutlineUsers />, href: '/dashboard' },
+  { title: 'Employees', icon: <HiOutlineUsers />, href: '/dashboard' },
   { title: 'Credits', icon: <HiOutlineCurrencyDollar />, href: '/credits' }
 ]
