+++ apps/front/common/routes.ts
@@ -1,0 +1,3 @@
+export const ROUTES = {
+  DASHBOARD_PATH: '/dashboard'
+}
