+++ apps/front/pages/dashboard.tsx
@@ -2,9 +2,9 @@ import { InferGetServerSidePropsType } from 'next'
 
 import { bungeeHairline } from '@/common/styles/fonts'
 import { DashboardLayout } from '@/common/Layout/ua/DashboardLayout'
-import { PeopleList } from '@/modules/dashboard/components/PeopleList'
 import { getUserSession } from '@/common/utils/getAdminSession'
 import { useFetchCompany } from '@/modules/dashboard/hooks/useFetchCompany'
+import { EmployeeList } from '@/modules/dashboard/components/EmployeeList'
 
 export const AdminDashboardPage = ({
   user
@@ -16,8 +16,10 @@ export const AdminDashboardPage = ({
       nameInitial={user?.firstName ? user.firstName[0] : ''}
       token={user?.access_token}
     >
-      <p className={`pt-6 text-3xl ${bungeeHairline.className}`}>People List</p>
-      <PeopleList
+      <p className={`pt-6 text-3xl ${bungeeHairline.className}`}>
+        Employee List
+      </p>
+      <EmployeeList
         employees={companyData?.data.employees}
         isLoading={isCompanyLoading}
       />
